+++ v3/.example/config/observability.go
@@ -1,12 +1,8 @@
 package config
 
 import (
-    nethttp "net/http"
-
     melodyopentelemetry "github.com/precision-soft/melody/integrations/opentelemetry/v3"
     "github.com/precision-soft/melody/v3/exception"
-    melodyhttpcontract "github.com/precision-soft/melody/v3/http/contract"
-    melodyruntimecontract "github.com/precision-soft/melody/v3/runtime/contract"
 )
 
 func (instance *Module) buildObservability() {
@@ -17,12 +13,4 @@ func (instance *Module) buildObservability() {
 
     instance.metricsMiddleware = middleware
     instance.metricsHandler = handler
-}
-
-func metricsRouteHandler(handler nethttp.Handler) melodyhttpcontract.Handler {
-    return func(runtimeInstance melodyruntimecontract.Runtime, writer nethttp.ResponseWriter, request melodyhttpcontract.Request) (melodyhttpcontract.Response, error) {
-        handler.ServeHTTP(writer, request.HttpRequest())
-
-        return nil, nil
-    }
 }